requirements.txt
fastapi
uvicorn
genanki
google-generativeai
python-dotenv
pydantic
pydantic-settings
gTTS
duckduckgo-search
pypdf
numpy
requests
